chore(gcp): add provider
