+++ prowler/providers/gcp/services/apikeys/apikeys_service.py
@@ -1,12 +1,13 @@
 from pydantic import BaseModel
 
 from prowler.lib.logger import logger
+from prowler.providers.gcp.gcp_provider import GcpProvider
 from prowler.providers.gcp.lib.service.service import GCPService
 
 
 ################## API Keys
 class APIKeys(GCPService):
-    def __init__(self, provider):
+    def __init__(self, provider: GcpProvider):
         super().__init__(__class__.__name__, provider, api_version="v2")
 
         self.keys = []
