+++ prowler/providers/gcp/services/dataproc/dataproc_service.py
@@ -1,13 +1,14 @@
 from pydantic import BaseModel
 
 from prowler.lib.logger import logger
+from prowler.providers.gcp.gcp_provider import GcpProvider
 from prowler.providers.gcp.lib.service.service import GCPService
 from prowler.providers.gcp.services.compute.compute_client import compute_client
 
 
 ################## Dataproc
 class Dataproc(GCPService):
-    def __init__(self, provider):
+    def __init__(self, provider: GcpProvider):
         super().__init__(__class__.__name__, provider)
         self.regions = compute_client.regions
         self.clusters = []
